# docs/swagger.yaml
definitions:
  schema.Login:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  schema.Register:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 登录
      parameters:
      - description: 登录
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.Login'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 登录
      tags:
      - 认证
  /auth/register:
    post:
      consumes:
      - application/json
      description: 注册
      parameters:
      - description: 注册
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schema.Register'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 注册
      tags:
      - 认证
  /problems:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      - description: 关键字
        in: query
        name: keyword
        type: string
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取题目列表
      tags:
      - 公共方法
  /problems/{id}:
    get:
      parameters:
      - description: 题目ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: '{"code":200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取题目详情
      tags:
      - 公共方法
swagger: "2.0"